Impl faction augs filter helper function
